rename `floatob.rs` to `float.rs`
